refactor: Appliquer le pattern Server/Client Component et intégrer l'AppBar
- Refactorisation de la page login selon le pattern Server Component + Client Component
  - Server Component vérifie l'auth et redirige si déjà connecté
  - Client Component gère toute la logique interactive du formulaire
- Création du layout protégé avec AppBar fixe en haut de l'écran
- AppBar maintenant sticky avec position fixed et z-index élevé
- Suppression de la page signup (non utilisée dans un contexte mono-tenant)
- Amélioration des performances avec vérification auth côté serveur
