+++ app/(auth)/login/page.tsx
@@ -1,110 +1,21 @@
-"use client";
+import { redirect } from "next/navigation";
+import { createClient } from "@/lib/supabase/server";
+import { LoginPageClient } from "./login-page-client";
 
-import { useState } from "react";
-import { useAuth } from "@/lib/hooks/use-auth";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { Alert, AlertDescription } from "@/components/ui/alert";
-import { Loader2, Mail } from "lucide-react";
-import Link from "next/link";
+export default async function LoginPage() {
+  // Server-side authentication check
+  const supabase = await createClient();
 
-export default function LoginPage() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState("");
+  // Check if user is already authenticated
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
 
-  const { signIn } = useAuth();
+  // Redirect to dashboard if already logged in
+  if (user) {
+    redirect("/dashboard");
+  }
 
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-    setError("");
-
-    try {
-      await signIn(email, password);
-    } catch (err) {
-      setError(
-        err instanceof Error ? err.message : "Une erreur s'est produite"
-      );
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <div className="flex min-h-screen items-center justify-center bg-gray-50 px-4 py-12 sm:px-6 lg:px-8">
-      <Card className="w-full max-w-md">
-        <CardHeader className="space-y-1">
-          <div className="mb-4 flex items-center justify-center">
-            <Mail className="text-primary h-8 w-8" />
-          </div>
-          <CardTitle className="text-center text-2xl">TrackedMail</CardTitle>
-          <CardDescription className="text-center">
-            Connectez-vous à votre compte pour accéder au tableau de bord
-          </CardDescription>
-        </CardHeader>
-        <form onSubmit={handleSubmit}>
-          <CardContent className="space-y-4">
-            {error && (
-              <Alert variant="destructive">
-                <AlertDescription>{error}</AlertDescription>
-              </Alert>
-            )}
-            <div className="space-y-2">
-              <Label htmlFor="email">Adresse e-mail</Label>
-              <Input
-                id="email"
-                type="email"
-                placeholder="votre@email.com"
-                value={email}
-                onChange={e => setEmail(e.target.value)}
-                required
-                disabled={isLoading}
-              />
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="password">Mot de passe</Label>
-              <Input
-                id="password"
-                type="password"
-                placeholder="••••••••"
-                value={password}
-                onChange={e => setPassword(e.target.value)}
-                required
-                disabled={isLoading}
-              />
-            </div>
-          </CardContent>
-          <CardFooter className="flex flex-col space-y-4">
-            <Button type="submit" className="w-full" disabled={isLoading}>
-              {isLoading ? (
-                <>
-                  <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                  Connexion en cours...
-                </>
-              ) : (
-                "Se connecter"
-              )}
-            </Button>
-            <p className="text-muted-foreground text-center text-sm">
-              Pas encore de compte ?{" "}
-              <Link href="/signup" className="text-primary hover:underline">
-                Créer un compte
-              </Link>
-            </p>
-          </CardFooter>
-        </form>
-      </Card>
-    </div>
-  );
+  // Render login form if not authenticated
+  return <LoginPageClient />;
 }
