+++ components/app-bar.tsx
@@ -27,7 +27,7 @@ const navigationLinks = [
 
 export function AppBar() {
   return (
-    <header className="border-b px-4 md:px-6">
+    <header className="bg-background fixed top-0 right-0 left-0 z-50 border-b px-4 md:px-6">
       <div className="flex h-16 items-center justify-between gap-4">
         {/* Left side */}
         <div className="flex items-center gap-2">
